fix(libpfm): problem with configuration, fix #41
Change the list conversion to be able to read the variables from a configuration file.

fix(cpu-sensor): problem with the TargetUsageRatio when the All target is received, fix #42

Create a case class to store the global cpu time (with idle) and the active global cpu time (without idle) in order to compute correctly the TargetUsageRatio.

refactor(PowerMeter): improve code

Use the configuration trait instead of an object with a default timeout, change the null by Option.