refactor(LibpfmHelper): adds methods to detects the PMU, the number of generic counters and the events available on the targeted CPU
We are now able to detect automatically the PMU with their number of counters that we can open simultaneously.
This number is very useful to know exactly how many events that we can monitored at a the same time.

feature(Libpfm): creates a generic formula for Libpfm and builds modules upon this generic formula

This formula can be configured with a map of coefficients for each event which forms the formula.
It represents the entire processor's consumption with several events.