feature(code-energy-footprint): allows to get the power consumption of a piece of code (represented as a label) from an external program
The method for configuring a PC is now more configurable and easier to use in different scenarios.
Different picker is now available for creating or using file descriptors upon needs.
Creates a UnixSocket server to handle the connexion with clients, to receive information and file descriptors from outside.
Creates a new sensor for creating the server, for processing information sent by clients and for getting performance counter values from codes (represented as labels).
Updates the CLI for monitoring a piece of code (+ improvement).

feature(code-energy-analysis): adds an application for analyzing the traces of an instrumented program in order to obtain the power consumption of each call through the time.

Creates an application to gather the power estimation of PowerAPI and the traces produced by an instrumented program.
The traces are processed to a global call graph, and the power consumption are then correlated to time spent per call.