refactor(core): add Supervisors, fix bugs, cleaning, a Clock is not automatically created anymore
A Clock for producing periodic ticks is created only when the "every(frequency)" method is called on a PowerMonitoring object.
It allows then to have external ticks when needed.
Sensors, formulae and reporters have now their own Supervisor, and each component actor is now responsible to handle one target at a time.
Two reporters are not refactored yet (Rest, Fuse).

refactor(tests): improve the coverage, and clean tests

Several tests have been rewritten and cleaned.
The coverage is now up to more than 70%.

refactor(examples): change the external dependency to a local one

The examples were written to use PowerAPI as en external dependency. The dependency is now local (PowerAPIBuild), and allows to get the examples up-to-date with the powerapi-core library.