feature(code-energy-analysis): create an app for producing power estimation at method level
The new app acts like a daemon for getting information about the function which has been at the origin of the interruption.
A PowerAPI agent has to be attached to a program (which can be compiled with the -rdynamic option for now) for setting the interruption mode on it and getting information about the running function when the handler is triggered.
The agent then sends data to PowerAPI other Unix sockets for computing power estimation.